Add CMakeLists.txt for C code
Split into shared library and main command line executable.
